+++ changelog.md
@@ -1,4 +1,4 @@
 Changelog next version:
 
 * Catchup subscriptions (blocking)
-* EveryN for stream persistence
+* EveryN for stream persistence (both blocking and reactive)